CDK/lambdapipeline/requirements.txt
aws_cdk.aws_s3
aws_cdk.aws_iam
aws_cdk.aws_lambda
aws_cdk.aws_dynamodb
aws_cdk.aws_sqs
aws_cdk.aws_logs
aws_cdk.aws_s3_notifications
boto3
moto